+ refactor settings       -	Make settingsBase abstract       -	Add AudioSettings
git-svn-id: svn://svn.code.sf.net/p/droidcachebox/code/trunk@1780 5890f383-b09f-485f-8f59-22bc967d67fd

+++ CB_UI/src/CB_Core/GL_UI/Views/MapView.java
@@ -333,7 +333,7 @@ public class MapView extends CB_View_Base implements SelectedCacheEvent, Positio
 		drawingWidth = mapIntWidth;
 		drawingHeight = mapIntHeight;
 
-		iconFactor = (float) Config.settings.MapViewDPIFaktor.getValue();
+		iconFactor = Config.settings.MapViewDPIFaktor.getValue();
 
 		togBtn = new MultiToggleButton(GL_UISizes.Toggle, "toggle");
